skin-lesion-segmentation/requirements.txt
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
monai==1.4.0
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.11.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
